+++ tobbo.kirill/I2/methods.h
@@ -3,6 +3,6 @@
 #include <cstddef>
 
 void count_of_maxes(const int* arr, size_t size, int& maximum, size_t& count_maximum);
-void shift_part_to_end(int* arr, size_t size, unsigned int m, unsigned int n);
+void shift_part_to_end(int* arr, size_t size, size_t m, size_t n);
 
 #endif
